requirements.txt
pyyaml
tqdm
pygame
pydantic